requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
sqlparse==0.5.1
urllib3==2.2.3
